+++ .eslintignore
@@ -1,3 +1,5 @@
 /dist
 /examples
 tsconfig.json
+esbuild.config.js
+*jest*
