+++ examples/node/esm/index.mjs
@@ -1,3 +1,3 @@
-import {cep, Provider, factory, CepService, Requester} from '../../../dist/esm/index.mjs'
-console.log(cep, Provider, factory, CepService, Requester)
+import {cep, Provider, factory, CepService} from '../../../dist/esm.mjs'
+console.log(cep, Provider, factory, CepService)
 cep('41342315').then(console.log)
