+++ jest.config.cjs
@@ -1,0 +1,6 @@
+module.exports = {
+	testEnvironment: 'node',
+	testMatch: ['**/?(*.)+(spec|test).[jt]s?(x)', '**/?(*.)+(spec|test).mjs'],
+}
+
+
