--- src/cep.ts
@@ -1,8 +1,0 @@
-import { Cep } from './entity/'
-import factory from './factory'
-export const cep = (cep: string): Promise<Cep> => {
-	const handler = factory({
-		useDefaultProviders: true,
-	})
-	return handler.execute(cep)
-}
