--- src/entity/cep.ts
@@ -1,7 +1,0 @@
-export type Cep = {
-	cep: string,
-	street: string,
-	state: string,
-	neighborhood: string
-	city: string
-}
