--- src/entity/index.ts
@@ -1,3 +1,0 @@
-import { Cep } from './cep'
-
-export { Cep }
