--- src/index.ts
@@ -1,14 +1,0 @@
-import { CepService } from './service/index'
-import { cep } from './cep'
-import factory from './factory'
-import { Provider } from './provider'
-import { RequestWIthFetch, Request } from './requester/index'
-
-export {
-	cep,
-	CepService,
-	Provider,
-	factory,
-	RequestWIthFetch,
-	Request
-}
