+++ src/service/brasilAPI/index.mjs
@@ -1,10 +1,8 @@
-import { Request } from '../../requester/index'
-import { CepService } from '../index'
+import { CepService } from '../index.mjs'
 
 export class BrasilAPIService extends CepService {
-	constructor(request: Request) {
-		super('brasilAPI', request)
-		this.baseUrl = 'https://brasilapi.com.br/api/cep/v1'
+	constructor(request) {
+		super('brasilAPI', request, 'https://brasilapi.com.br/api/cep/v1')
 	}
 
 	async handler(cep) {
